requirements.txt
# Email Parsing
extract-msg==0.47.0
olefile==0.47

# Natural Language Processing
scikit-learn==1.3.2
nltk==3.8.1
spacy==3.7.2
yake==0.4.8

# AI/LLM Integration
anthropic==0.18.1

# Document Generation
python-docx==1.1.0

# Data Processing
pandas==2.1.4
numpy==1.26.2

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Utilities
tqdm==4.66.1
colorama==0.4.6

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Development
black==23.12.1
flake8==7.0.0
